requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.20.0
matplotlib>=3.5.0
tqdm>=4.65.0
llama-cpp-python>=0.2.0
matplotlib-font-manager>=0.1.0